Fix problems with "typedef struct foo_s foo;" redefinitions.
In the header revamp, I'd left various places where we
use the above formulation in more than one location. This
is fine with all the C compilers I tested it on, but seems
to update gcc 4.1.2.

This effect can be seen on later gcc's by using
"-std=gnu89 -pedantic".

This commit tweaks the code to avoid redefinitions of this
kind.