Fix crash reported by customer #531 with multi-threaded rendering.
This was a race condition where threads were continuing to be started
while we were tearing threads down. This doesn't occur with most usages
because we render the entire page, but with the customer's device, it
stops reading before the last line, but the multi-threading rendering
logic tried to keep starting threads (up to NumRenderingThreads) to
have later bands available.

The solution is to wait for all threads to be in the "THREAD_DONE"
state before freeing the threads.

Also, since the color_usage_array of the main thread is shared by the
rendering threads, wait until the threads are shut down before freeing
that.