fix some false positive circular reference detections
If we had two 'Box' entries in a page dictionary, and both were indirect
references to the same matrix, we would incorrectly spot them as a
circular reference.

This fixes the problem.

I have a notion for a better way to deal with circular references but
it will be disruptive, and I'm not certain its better. I'll experiment
later.