Fix some errors spotted by valgrind.
The buffer we were supplying to sscanf wasn't null terminated. Oddly
the same buffer didn't trigger a valgrind error on strstr, I have no
idea why. Fix it by reading one less byte and null terminating the
string.

Also ensure we read enough bytes to satisfy the header detection; we
need 8 bytes for %PDF-X.x. Prevents a potential buffer overrun in the
sscanf pointer.

Finally, the detection of line endings was incorrect when scanning for
startxref, we were using the number of bytes read (which is a 1-based
index) to index an array of chars which is of course a 0-based index.

This gets rid of 2 valgrind errors.