Fix (many) indetermisms in cluster testing.
Previously when we passed in a pointer for a returned colour space the
colour space creation code did not set the space. This turned out to
be a problem for images and 'unusual' spaces like Separation or DeviceN
because the space was not properly installed.

For now, set the colour space even if we are returning it. We may need
to change this in future, I'm not sure we always will want to do this.

Also, do a gsave/grestore around the image drawing, so that we don't
change the colour space unexpectedly. This should probably have been
done anyway so that Form XObjects don't change the graphics state.