Fix psdcmyk devices
This commit comprises a number of individual fixes, splitting them up
would result in cluster runs which have large numbers of seg faults or
garbage output so I've chosen to address them all in one commit.

The main change is to the spot colour detection. The existing PDF
interpreter runs multiple passes to detect spot colours and transparency
which require us to walk the Resources trees multiple times. Currently
we first walk all the Resources in all the pages to determine whether
there are any spot colours. We then process each page individually and
detect whether any Resource in that page uses transparent operations.

Initially I was trying to do the checking on a per page basis, but that
doesn't work well for spot colour detection. So the major change in this
commit is to move the transparency and spot colour detection to an
initial scan when the file is opened. To do this we need to be able to
store the transparency state of each page (because the PDF14 compositor
is set/reset on each page). So there's a new array of bits to track the
transparency state of each page 'PageTransparencyArray'. We could store
the state in the page dictionary, but that would mean storing an
additional Key and Value per page, which is considerably more than one
bit.

To reduce the overheads, we only check for spot colours if the output
device can accept a declaration of Spots up front (ie psdcmyk). To keep
track of this, store the device capability in a new boolean in the context.

We also short-circuit the transaprency detection if we either have set
-dNOTRANSPARENCY, or we have already detected transparency.

The pdfi_check_ColorSpace_dict function was incorrectly insisting that
The entries (after the first) in a /Resources /ColorSpace dictionary
should be dictionaries, they must in fact be arrays.

The change to remove PDF_INDIRECT checks, and have pdfi_dict_get always
dereference broke the Pages array processing. It actually worked
correctly the first time we retrieved any given page (which is why the
breakage wasn't noticed) but failed the second time we tried to get a
page dictionary, because we expected to get an indirect object returned
from a call to pdfi_dict_get, but actually got the dereferenced object.

This also led to teh interesting discovery that one of the Quality Logic
test files (FTS_07_0704.pdf) has an invalid page dictionary; instead of
the dictionary containing a key with the value /Page, it has a key with
the value /Template. Acrobat (of course) silently ignores this. So
rejig the Pages array processing to ignore incorrect /Type entries in
Page dictionaries, but flag the error and warn on exit.

Taken together this results in a very large number of progressions,
mainly with the psdcmyk device. The repeated bands, coloured noise
and vertical lines in the output are resolved for almost all input tests
(but see below). Additionally there are some colour shifts; mostly these
are too subtle to detect but where the difference is visible these seem
to be generally progressions.

There is one known problem with this commit, a /Indexed colour space
whose base space is /Separation or DeviceN does not detect that spot
colour. If this is the only use of the spot in the file then the initial
scan won't set the number of spots correctly, leading to the same kind
of problems fixed above with the psdcmyk device. It also (I think)
leads to at least some seg faults.

I know the problem exists and I am working on a fix but I want to commit
it separately in order to check the stability of the cluster runs.