Bug 700952: re-introduce over/underflow workaround
Commit 355434f4b1bbe8c4f98cafad5a6868aa2f0eaae1 reverted a workaround that
compensated for over/underflow in Freetype's TTF hinting (related to freedom
and projection vector calculations). That problem no longer exists in recent
Freetype releases, and the workaround actually caused other issues to occur
with hinting.

What wasn't obvious was that the workaround also protected over/underflow
issues relating to the unitsPerEm value.

So this re-instates the workaround, but bases the decision on how the final
scale is distributing between the Freetype "size" and the Freetype matrix on
the unitsPerEm value (this is relevant for all font types as, for non-TTF,
font types, Freetype derives the unitsPerEm from the FontMatrix (for PS type
fonts).

Also fixes Bug 700875