Change s_jbig2decode_release() to free the global context in some cases
Basically, if state->global_struct is NULL, it means the library should
free the global_ctx when it's done with it.  This is the case when used in
gpdf (C implementation of PDF).

If the state->global_struct is non-NULL, then it is a memory-managed
value that the interpreter will eventually clean up.  This is the case
when used in gs (PS implementation of PDF).