Bug 700968: Fix TTF Format 4 (segmented) cmap parsing
The format 4 parsing was incorrect, but in a manner only exposed by reading
an out of spec cmap.

The spec implies that the character codes in the cmap segments should be non-
overlapping, and the segments should be in order of lowest-to-highest codes.

We were imposing (*not* enforcing) that condition, but forcing the start code
of the current segment to be greater than size of the range of codes we've
already read. This, basically, has no effect on fonts that adhere correctly
to the spec.

The fonts embedded in the files for Bug 700968 do not conform to the spec. For
example, the 'Z_C00002.tmp' font has the first 92 segments in the cmap with
start code and end code both set to 33. The above imposing of the no-overlap
and order of segments meant that when we got past those first 92 segments, we
were forcibly using character codes 92 and upwards.

Removing the code imposing those conditions causes these files to render
correctly, and shows no regressions in our existing test suite.