PDF interpreter - synthesise appearance for text Widget comb fields
Bug #700967 "Warning Combed form fields are not yet implemented"

Comb fields draw a string into a series of 'combs' or boxes, one
character per box. As usual there are no guidelines in the specification
as to how to draw the individual characters.

I've chosen to centre each character in the box it is draw in, because
drawing it from the origin of the box tends to cause collisions with the
vertical/horizontal lines delineating the box itself (in the one example
file I have). This may cause slightly odd appearance if a field uses
mixed upper/lower case, as the text will jump up and down vertically.

The example file supplied has a fault, one of the comb fields has a
value (/V) string which is longer than the /MaxLen value for the field.
This indicates to me a fault with the form filling application which
should not permit this. Acrobat truncates the string, I've chosen to
not render the string at all, and emit a warning on the back channel.

None of our test suite files show any differences at all; either we
don't have an examples of this kind of field already, or they all have
appearance streams.