Documentation - describe the ink_cov device
Following up from a question on Stack Overflow, the ink_cov device isn't
actually described anywhere. Add a description of the device and how its
output differs from the inkcov device.

At the same time, add the inkcov device(s) to the table of contents in
devices.htm and also add the bit and bbox devices which were not listed
in the TOC.