Bug 700624: Fix "full" pattern fills with gaps between cells.
GS has an optimisation in it, whereby if an uncolored pattern fill
is completely full (the mask is full of 1's), it doesn't bother to
store the mask. When such a pattern is filled, the code knows just
to fill the whole thing.

Sadly, this optimisation breaks down when the cells have an X or Y
step that means there should be gaps between the instances. GS
currently fills the whole thing, gaps and all.

To fix this, we either need to fix all the different fill routines
to cope with this case, or (as I have done here), disable the
optimisation in the case when the X and Y step is larger than the
'size' of the pattern cell.