PDF interpreter - change detection of CFF CIDFont types
Bug #700982 "Error: unrecognized font file starts with <01000D02>"

The code in pdf_font.ps to detect font types, and call an appropriate
font loader works (in part) by reading the first 4 bytes from the font
stream, and then using that as the key for a dictionary where the
values are the font loading procedures.

Most font types have a 'magic number' or similar at the start of the
font data, but CFF fonts do not. In fact the header of a CFF font is
highly variable, the only thing we can rely on being the font version
byte (0x01), the remaining 3 bytes are variable.

We've tackled this previously by simply adding more different varieties
of string values, but this seems nonsensical. There are too many
possible values for this to work reliably.

Ideally I'd recode this to use the CIDFont Subtype to select the
correct font loader, but that would mean recoding all of this, which is
more work than I want to tackle in this code. Instead I've chosen to
only have one single 'tag' in the dictionary, and I've extended the code
handling the case where we don't find a tag in the dictionary.

Previously this checked to see if the first two bytes of the string
were <8001> and otherwise failed. I've modified this so that if the
bytes are not <8001> then we check the dictionary /Subtype key. If its
not present or isn't either CIDFontType0C or 1C then we still throw
an error. Otherwise we discard the string we read, replace it with the
one and only string we have stored in the font_tag_dict dictionary and
proceed.

If there is ever a version 2 of the CFF specification this will fail,
but we can add a check for the version number if that ever happens
(seems highly unlikely).