Fix commit d89bae568d5015144e5d68fea8221543111e119e
Commit d89bae568d5015144e5d68fea8221543111e119e contained an oversight
which unfortunately was missed due to the testing cluster being in a
state of flux.

Its possible for fonts to have a Subtype of Type1C as well as the
noted CIDFontType0C and CIDFontType1C, and we need to cater for that
as well.

This mostly affected some of the Ghent Working Group test files, and
not much else, so I've chosen just to push an addditional commit.