fix /Pattern indeterminisms
Because the setting of the current colour space to be /Pattern is
currently disabled we can't rely on testing currentcolorspace in
pdfi_setcolorN to see if we're expecting a Pattern space.

So we need to check specifically for a name argument, assume its for
an (as yet not set) Pattern space and just ignore it.

While we're here, move the checking of the number of available arguments
so that we can check against the expected number, from the current
colour space (previously we only checked for at least 1).

I believe this fixes the indeterminate output.