Use new device colour space creation routines
We created the routines in a previous commit, but only used them in
some cases (for testing). Extend that to use them in all cases.

These routines handle 'ppcs' so that they either set the newly created
space or return it as requested. This will be extended to the remaining
(more complex) spaces in follow up commits but I want to do this piece
meal in order to isolate any problems.