Adjust a colour space reference and some code cleanup
The code in pdfi_seticc_cal() (copied from the PostScript interpreter)
doesn't count up a colour space if it is found in the cache. This is
obviously wrong, as whatever takes the reference will expect to be
able to count it down when its finished with it.

This requires us to decrement the reference count when setting an Lab
CalGray or CalRGB space.

Modify pdfi_crete_indexed() to handle whether its setting or simply
creating and returning a colour space (whether ppcs is NULL). Simplify
the use in pdfi_create_colorspace_by_array(). More spaces still to
cover.