Add support for named color when going to separation devicen_icc_cmyk
When going through named color processing with the target device either
tiffsep or psdcmyk, cmap_devicen_direct and cmap_devicen_halftoned were
doing an additional cmyk icc mapping.  This mapping was needed to ensure
that CMYK values matched DeviceN mapping values of CMYK+Spot where the Spot
value was 0 percent.  With the use of named color mapping, we want to avoid this
icc mapping.  Fixing this issue involved adding a boolean value to the DeviceN and Separation
color space params to indicate if the spot colorant(s) are supported by the
named color profile.  This determination is made during the install of the
device so that it need only be done once.  To do that, we needed to add a
method, gsicc_support_named_color, in gsicc_cache.c, which goes with
gsicc_transform_named_color, to indicate if the spot colorant is defined in
the data that we are using.  Tested with tiff24, tiff32 and psdcmyk to ensure
named color mapping is occurring and that no additional mapping occurs after
that.