pdfwrite - handle user-defined PCL fonts with character codes > 256
Bug #700996 "PCL Files Processing failure"

The problem is that we cannot handle a character code in excess of 256
in a PDF Font (we would have to create a CIDFont, which is too much
work). This was causing us to fail to attach a PDF font to the current
(user-defined PCL) font. When we later tried to retrieve the font
that caused a seg fault.

If we simply return an error when there is no attached font then the
fallback machinery will take us through generating an image instead.

This, however, caused the original file not to work properly, text was
going missing. Forcing 'early accumulator' to be 0 so that we created a
new font resolved the problem. Its better, however to remove that entire
section of code, since it now has no purpose.

This commit results in a PDF file with the correct appearance. A number
of out test suite files exhibit small differences, particularly at
low (75 dpi) resolution, but nothing is actually incorrect.