Save default_ctm, to use it with patterns.
Saving the value from the beginning of the page, then using that value
when setting a pattern.

It's possible I need to save/restore the whole gstate and not just the
CTM, but just using the CTM fixed a LOT of stuff.  This is a WIP...

Wrote some "null" pattern initialization code, but it segfaults a few samples
and doesn't seem to actually be needed at the moment.  Left it in, disabled.

This commit has quite a few changes in bmpcmp, but the 20 or so that I looked
at all seemed to be progressions, so overall it seems like a good change.