Bug 700973: CIDFont substitution: avoid multiply applying metrics
findCIDFont: we make various efforts to produce a synthetic font style
(bold, light, italic etc) if the actual type face style is not available.

In these cases, we have to avoid reusing an existing CIDFont substitute because
the CDevProc will have been 'hooked' to apply the width metrics drawn from the
PDF file. We want to, if at all possible, use the CIDFont as originally loaded.