Have CMYK colors in DeviceN color spaces color managed when going to CMYK devices
The current code was only doing color management of DeviceN CMYK when going to
the separation device (e.g. tiffsep and psdcmyk).  It should also be doing this
for pure CMYK devices like tiff32nc etc.   With RGB devices we will be going through
the alternate tint transform.  The would only pertain to DeviceN color spaces composed
of only process colors (e.g. [Cyan Magenta]).  If a spot color was included we would
be going through the alternate tint transform for a device like tiff32nc.

So with this commit if you have a command line like

-sDEVICE=tiff32nc -sDefaultCMYKProfile=src_cmyk.icc -o output.tif -f input.pdf

it will treat the CMYK values in the DeviceN color space the same as a DeviceCMYK
source color and make use of the src_cmyk.icc profile to handle those.  As mentioned
this worked with psdcmyk and tiffsep but not standard CMYK devices.