Add methods to detect pure process colors in Sep and DeviceN color spaces
When the colors in the separation or Devicen color space consist all of
one particular process color model, we would like to consider the use
of the alternate tint transform color space as an indication of the intent
for these color spaces (assuming the attributes information was not there).
For example, if we had a DeviceN color space
of [Cyan Magenta] and the alternate tint transform color space was a
CMYK ICC color space, that would be a hint that we should treat the color
as being in that space.  This commit provides us with the information as
to the colorants all being in one particular process space which is
a requirement.