Process DeviceN attributes dictionary, /Process dictionary, /ColorSpace
PostScript PDF interpreter changes; retrieve the Process dictionary
from the DeviceN attributes dictionary (if present). Extract the
ColorSpace, resolve the ColorSpace (turns any references into actual
objects), and store the result back in the Process dictionary. Finally
store the Process dictionary in the attributes dictionary which we
pass to setcolorspace.

PostScript interpreter, C changes; add a new 'stage' for DeviceN
colour space handling. In stage 3, check for the existence of an
attributes dictionary, then a /Process dictionary, and a /ColorSpace
entry. Retrieve the entry. Set up a new continuation procedure
'devicenprocess_cont'. The continuation procedure calls zsetcolorspace()
to set the /ColorSpace as the current space. On returning, retrieve
the current color space, attach it to the DeviceN graphics state colour
space and finish.

Still to do; process /Components, add the equivalent code to pdfi,
check pdfwrite to see if we can handle this. Tidy up comments.
Committing now so Michael can carry on with the CMS work.