Add holder for process color names in DeviceN color space parameters
As part of the work to use the DeviceN attributes process dictionary
we need to have a place to store the colorant names.  They will
be stored as an array of C strings in the pcs->params.device_n structure.
This commit adds them to the structure and frees them during the color
space finalization.