Type 1 patterns working (for some definition of working).
We use the pattern rendering device that caller has setup, and render into
it by interpreting the pattern stream.

This version doesn't try to handle high level devices, transparency,
probably some other special cases.

---
Also: Fix new crash on Bug696410.pdf

This appears to be a corrupted PDF (see bug report).
It won't work, but but now it won't crash.
It is trying to load a pattern while rendering a pattern, and getting
weird results.  Added a check for that pattern type to be a dictionary
and now return an error instead of crashing.