Use deviceN attributes process color color space
If the DeviceN color space includes an attributes entry and
it has a process color space, and the color space is of the proper
type use that as the source color for the process components.  If
instead the alternate tint transform color space is the correct type
use that as a hint.  If neither work then set to the defined default_cmyk
source color space.  If all fail due to issues with ICC profiles then
the CMYK source process colorants will pass through to the output device.