High-level device patterns (type 1 pdfwrite)
This implementation works without needing to return back out for a
remap (no gs_error_Remap_error needed), but requires a small change to
the graphics library for the gs_error_handled error that is now
returned for this case.

Normalizes and adjusts the bbox rectangle. Fixes the code to do the
same hackery that the ps implementation does.

Tune handling of errors during pattern stream (see Bug692245.ps).
Need to have pattern sub-stream stop on error, and also put in code
to clean up the stack if needed.