DeviceN - handle attributes Subtype and cleaning up
Previously we were ignoring the /Subtype in the attributes dictionary
of a DeviceN colour space (if attributes were present). We may need to
be able to tell the difference (now that we handle /Process) because
the values associated with the process components must be stored in
their natural form (eg subtractive for CMYK, additive for RGB) when
the Subtype is NChannel.

The remainder of the commit tidies up naming; previously we had a
structure called gs_devicen_attributes, which was used to store the
Colorants entry form an attributes dictionary. Since we now handle
other aspects of the attributes dictionary as well, this was misleading
so the structure has been renamed throughout as gs_devicen_colorant
(since each structure defines only one colorant, the plural name was
also misleading). Apparently pdfwrite is the only client for this
currently anyway.

Finally, moved the 'devn_process_space' from the general graphics
state colour space structure into the DeviceN params structure, as its
only relevant to DeviceN (and indeed only a DeviceN with an attributes
dictionary which contains a /Process entry)