DeviceN - store C strings for Colorant names, not name indices
For DeviceN spaces with an attributes dictionary containing a
Colorants dictionary we were storing the Colorant names as a name index
(index into the interpreter name table). Additionally we store a
procedure callback to recover the C string from the name index.

This is problematic for several reasons; firstly it requires all the
interpreters to either have a name table or some close equivalent so
that a C string can be recovered from a numeric value.

Secondly the lifetime of the string is divorced from the lifetime of the
C structure using it. Because there's no easy way for the interpreter to
know when the graphics library has freed the colour space structure
the interpreter is forced to retain the relationship between the name
index and the C string until the job is completed.

Both of these will be problems for the new PDF interpreter, so this is
step 1 in altering this. After this we will still need to change the
storage of the DeviceN and Separation ink names, which are handled the
same way.