DeviceN - change ink names to be C strings instead of name indices
As indicated in previous commits, change the stored ink names in the
colour space structure params.device_n from numeric values
(gs_separation_name) into C strings.

We never have to pass these back to the interpreter, so we allocate the
storage using the non-GC allocater, which simplifies things for the
array which used to store the indices and now stores the pointers to the
C strings.