PostScript colour space handling
We were not catering for a /Subtype in a DeviceN attributes dictionary
having a value which is a string type not a name type. I can't quite
see  how we are ending up with a string type here, since the PDF file
has it as a name, but it does. Dealing with both resolves many errors.

The second change is more far-reaching; when processing an attributes
dictionary for a DeviceN colour space we may need to create colour
spaces for the Process dictionary /ColorSpace and for any of the entries
in the Colorants dictionary. We were doing that by calling
zsetcolorspace() which mostly works well, however there is one
optimisation which can cause a problem; if the target space is the same
as the current *interpreter* colour space, then zsetcolorspace()
simply returns.

The problem is that if we are halfway through processing a DeviceN
space then we haven't set the interpreter space to the new colour space
array yet which can mean that we incorrectly detect one of the target
spaces as being the same as the current space. This leads to us trying
to use the current space in the graphics state (the DeviceN space) as
the target space, with unfortunate consequences.

So create a new version of zsetcolorspace which doesn't do that
optimisation and use that when processing the various target spaces in
the DeviceN attributes dictionary.