Make sure transparency page group is isolated
Per the spec, the page group should be treated as an isolated
group.  This allows the page group to specify a blending color space.
This resulted in over 90 pages of bmpcmp diffs.  I went over all
of them.  They are progressions.  Most are very minor color differences
as we end up doing the blending and drawings in different color spaces
compared to prior to the commit