Add file access control to the 'gp_stat' method.
This is complicated by the need to add the memory allocator context to the
gp_stat parameters. To facilitate this, I've added a gs_memory_t pointer to the
gx_io_device structure, which is set during the io device initialisation at
startup.

mkromfs: Add dummy gp_stat_impl for Windows - saves quite a bit of pointless
build upheaval