Improve test for PS resources in romfs
Previously, although we always built a romfs, building gs with COMPILE_INITS=0
meant we included a romfs which was devoid of content, and an attempt to find
the status of the device would return 'gs_error_unregistered'. We used this
to determine whether to automatically add '%rom%Resource/Init' to the default
search path.

In the gpdl case (for the sake of PCL and XPS), we always have a usable romfs
(even with COMPILE_INITS=0) for the ICC profiles, so the above test gave a
false positive: gs would add the romfs to the search path, despite there being
no usable Postscript resources in it.

So, revise the test to explicitly check for Resource/Init/gs_init.ps on the
romfs. If the romfs is, indeed, empty, it will still return
'gs_error_unregistered', but in the gpdl case, it will return
'gs_error_undefinedfilename', and we correctly skip adding the romfs path to
the search paths for Postscript resources