Add Colorants and Process dictionary processing to DeviceN
Needed to use the gs_attachcolorant() function in gsdevn.c in order
to be able to construct the structure required for Colorants. However
this function makes assumptions about how the colour spaces are stored
which is only true for the PostScript interpreter.

This commit adds a new function which is more general.