Change detection of invalid use of abbreviated filters to a warning
Create a new flag in the PDF context 'pdf_warnings' and a new enumerated
data type for it.

When processing a PDF file, the Pattern code wants to set stoponerror
to true, and capture the error, in case the pattern stream throws
any problems.

However, the file Bug693485.pdf then doesn't work, because the pattern
uses an image which is compressed using the A85 filter. This is illegal
outside of an inline image (though Acrobat, of course, ignores this).

Changing the inappropriate use of an inline filter to be a warning
instead of an error allows us to render the file properly. This is a
fairly clear-cut case where we can more or less guarantee that using
the abbreviated filter name won't actually cause a problem, so a
warning is reasonable.

We've had a PDFSTOPONWARNING for a while, but it had no actual use,
this commit is the first real use of it.