Have the ColorSpace code check for inappropriate use of abbreviations
There are some colour spaces (CMYK, G, I, RGB) which are only valid
when used with an inline image. If a PDF file uses these spaces outside
the context of an inline image we want to raise a warning, but carry on
and use the space.

This commit adds the 'inline_image' flag to the colour space creation
routine and propagates it to the routines used to create complex spaces
(eg Indexed, ICCBased, Separation and DeviceN). This allos us to
determine if the use is legal.

Have the image code pass in the flag appropriately when creating
spaces for images. Have the Pattern code always set it to false
(although we can fill an imagemask with a pattern, we can't do that in
an inline image).