Fix pdf/pdftop.c to correctly get rid of compiler warnings.
Revert change to pdf/pdftop.c (324fafa3bf2521e04b6ad9e41f53866a3b231345)
which actually caused everything to crash.

This code is crazy, as it is passing in a void * that happens to be actually
the value "1" if the parameter is a boolean.  Why???

Anyway, I got rid of the compiler warning.  Maybe there's a better way
to do all this?