PDF interpreter - ignore errors in some faulty FreeText annotations
Bug #701058 "Can not convert pdf to tiff"

The PDF file has a FreeText annotation with no appearance stream and
faults in the Annotation dictionary.

Firstly (not perhaps strictly a fault) the /Contents key has a string
value with nothing in it. This isn't exactly an error, but it makes the
annotation pointless.

More seriously the /DA key contains:

(//Helvetica 12 Tf 0 g)

The double // on the font name means we end up with an empty name
on the stack, which causes errors.

In this commit:

Execute the DA processing in a stopped context, so that if it fails we
don't throw an error.

Surround the DA processing with a mark so that we can clear the stack if
it leaves anything behind.

Surround all of that with a 'gstate...setgstate' or 'gstate...pop'. If
we get an error this allows us to restore the graphics state to the
point before we did the DA processing. If we didn't get an error then
we can just discard the saved gstate.

Finally, check the length of the Contents string. If its 0, don't even
try to draw it. If its more than 0, but less than 4, then don't attempt
to check for UTF16 (2 byte BOM + 2 bytes for character means a minimum
of 4 bytes).

This results in slightly different output from pdfwrite with one file,
tests_private/pdf/sumatra/1331_-_prints_garbled.pdf, which results in
very small (single pixel) differences at 72 dpi.