Handle trans group knockouts properly if page group is isolated
Page groups by definition have to be isolated.  A fix for this to
ensure that the proper group color space is used for the page group
revealed a problem when the source PDF actually indicated the page
group was isolated.  This led to an incorrect selection of subsequent
backdrops for any knockout groups contained.