pdfwrite - fix ColorConversionStrategy
No bug number, this arose from a question asked on Stack Overflow.

At some point I removed the enumerated value ccs_UseDeviceDependentColor
but I failed to remove the name "UseDeviceDependentColor" from the
list of strategy names. This meant that setting ColorConversionStrategy
to anything except LeaveColorUnchanged was 'out by 1' in the enumerated
data type.

This led (for example) to Gray being treated as RGB for the purposes
of the output colour space.