Add check for gs_error_handled for pattern handling
Some clients (specifically gpdfi) can now return gs_error_handled if
they already took care of everything in the PaintProc call.  In this
case, just do all the cleanup as for any other error, but it's not
really an error, so return error code 0.

This wll be needed for pdfi when merged in, should have no affect on current
code.