Fix bug 701045 - infinite loop in pcl parser.
The recent work to remove the specialized PL cursor code and replace
it with streams removed end of data detection PCL depended on.  End of
data is now detected by the PCL parser by requiring a minimum number
of bytes on the next call to the parser.  If the required number of
bytes are not provided the parser signals end of data by advancing the
read pointer to the end of the buffer.

The change precipitated some reorganization in how RTL is initialized,
as well.