Large commit to implement stack and gsave level hygeine around content streams
-This is a combination of 13 commits.
-This is the 1st commit message:

Adds 'save state' to the pdf_context which contains values stored before
interpreting the current stream. Each substream save these and updates
the ones in the pdf_context with the current values.

Have the page code use the saved context values to restore the stack and
gstate values at the end of the page.

Have the substream routine restore the stack and gstate at the end of
each substream, and rewind the underlying file to the saved position

Add a new function to determine how many objects are available on the
stack (taking into account the 'saved value)

Use the count_stack() routine in pdfi_pop() and don't pop objects
that are below the current bottom.

Use the saved gstate level in pdfi_grestore() and don't restore back
beyond the level at the start of the stream.

Add a new circular reference error flag and set that when we detect
a circular reference. Don't emit a warning for every occurence on
release builds

Add new warnings flags for too many Q's, too many q's, Garbage on the
stack, and stack underflow in a substream.

-This is the commit message #2:

add warning output, use pdfi_clearstack instead of manually flushing the
stack.

- This is the commit message #3:

Stack fixing

Use pdfi_count_stack() instead of subtracting stack pointers

Remove the operands on error condition.

Clear the stack if there were not enough operands supplied. Don't bother
if there was < 1 argument on the stack (its already clear)

Fix pdfi_build_function_2 so that it flags an error if the size of the
C0 and C1 arrays is not the same. Previously it could return 'no error'
even when the function hadn't been built.

Made pdfi_do_image_or_form non-static, for use by the SMask code in
pdf_gstate.c. Frankly I don't know how this avoided crashing before
because the function being called (pdfi_Do) expected to pull an operand
from the stack, and we were not supplying one.

Fix an oversight in previous commit, when resizing the stack, use the
*total* number of operands currently on the stack, otherwise we could
end up copying too few if we are in a sub stream.

Reading xref streams the code popped an object from the stack, but
there was never one there. If we'd checked the error returns from
pdfi_pop() this would have shown up before.

- This is the commit message #4:

If either of the Process and/or Colorants dictionaries are empty,
don't try to process them.

- This is the commit message #5:

Add some headers to solve compiler warnings

Move definition of pdfi_do_image_or_form to relevant header

- This is the commit message #6:

Initialise code to prevent a (correct) compiler warning

- This is the commit message #7:

Add an explicit crash after executing a stream if the save gstate
is not the same as the current gstate, or the saved stack count doesn't
match the current stack count. This is temporary and for debugging only!

Fix the 'Tr' operator, we were incorrectly claling the T* operator
function, which didn't consume the Tr operand.

- This is the commit message #8:

Fix an implicit cast. pdfi_array_size() returns an unsigned int,
subtracting 1 wraps around to a very large positive (64-bit) number
instead of -1.

This causes us to try and read an entry from an empty array.

- This is the commit message #9:

Really need to find the time to work one this. Its supposed to try and
find cases where a white space is missing between two PDF operators
or a PDF operator and another token.

For now, have it clear the stack, Bug690300.pdf has a QBT (ie Q BT)
and that leads to an unmatched 'q'. So we do need too try and address
this.

- This is the commit message #10:

literal string fix

If we get an escape character '\' then we check the following bytes, and
deal with the escaping as normal. If we don't recognise the character
following the escape as an escaped character, then we ignore it, this
is apparently not an error.....

File Bug691602.pdf has every space in the text strings preceded by an
escape character....

- This is the commit message #11:

Fix bug695872.pdf

The file has an inline image with /W and /H values which are real
numbers rather than integers. This is, obviously, crazy.....

Allow the image info code to read double precision real numbers and
cast them to an int rather than throwing an error.

This required the addition of a pdfi_dict_get_number2() function to
match the get_int2() function to read one of two possible key names.

- This is the commit message #12:

pdfi_begin_group - clean up on error

If pdfi_begin_group gets an error then we won't call
pdfi_end_transparency_group() which does a matching grestore for the
gsave in begin_group(). So if we get an error, we need to clean up
and restore the graphics state.

- This is the commit message #13:

This code to be removed totally later.

Add code to keep a count of transparency groups and force them to
balance around context streams.

Add a new warning, and warning text if groups don't balance (this can't
be a fault of the PDF file I think, that's us).